LayoutTests/editing/pasteboard/dataTransfer-types-pasting-plaintext-expected.txt

Tests that DataTransfer's types contains text/plain but not no image types when pasting a plain text. To manually test, click on the button below then trigger paste (Command+V or Control+V).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS types.includes("text/plain") is true
PASS types.find((type) => type.includes("png")) is undefined.
PASS types.find((type) => type.includes("jpeg")) is undefined.
PASS types.find((type) => type.includes("tiff")) is undefined.
PASS successfullyParsed is true

TEST COMPLETE
